Fix `np.sum` overflow on Windows
https://github.com/karpathy/nanoGPT/pull/270

+++ scripts/prepare_openwebtext.py
@@ -57,7 +57,7 @@ def prepare(
 
     # concatenate all the ids in each dataset into one large file we can use for training
     for split, dset in tokenized.items():
-        arr_len = np.sum(dset["len"])
+        arr_len = np.sum(dset["len"], dtype=np.uint64)
         filename = destination_path / f"{split}.bin"
         dtype = np.uint16  # (can do since enc.max_token_value == 50256 is < 2**16)
         arr = np.memmap(str(filename), dtype=dtype, mode="w+", shape=(arr_len,))
